APPLICANT UPDATE AGAIN
models.py
added checklist and feedback models

base profile.html
added an if and elif function so students cant submit if their status is submitted and can only submit if the status is returned

tprofile
updated the gender section to sex and added plm email to be seen inside the profile

shifter9 html
added checklist upload file button

views.py
added checklist.get to get files from checklist button

changed the creation process of faculty applicant user from faculty application form to worksheet form

reason: user will be created when the person clicked next in the first page of application after submitting requested details. the update changed that so that the account of the applicant will be created after clicking submit in the page.

updated tprofile and s profile

both now uses the first name and last name of the user to access the certain details of the person instead of the password as this proved to be buggy and sometimes un usable when the page was refreshed with this update the page will not crash.

also updated the file handling method where it will only submit the files if the variables had a value if none then it will show an error to the student. This update with the update in base profile html will create a security that the students cannot submit files into the db without a certain status